refactor: move back convert(Rows|Columns)ToData to data.convert.js
